requirements.txt
PyPDF2==3.0.1
python-docx==1.1.2
PyMuPDF==1.20.7
jupyter==1.0.0
matplotlib==3.10.0  # Optional for visualization
pytest==8.3.4       # Optional for testing
# Additional Dependencies

PyPDF2==3.0.1
python-docx==0.8.11

# Testing
pytest==7.2.0
unittest
